refactor(character): extract reusable CharacterCreateActions composable
Move the new-character / quick-create action card pair out of
CharacterListScreen into a standalone CharacterCreateActions composable so
it can be reused outside the list screen. No behaviour change.
