+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/character/presentation/component/CharacterListScreen.kt
@@ -1,35 +1,23 @@
 package com.cyrillrx.rpg.character.presentation.component
 
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Bolt
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.text.font.FontWeight
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -45,21 +33,14 @@ import com.cyrillrx.rpg.core.presentation.component.SimpleTopBar
 import com.cyrillrx.rpg.core.presentation.component.SwipeToDelete
 import com.cyrillrx.rpg.core.presentation.component.rememberOptimisticDeleteHandler
 import com.cyrillrx.rpg.core.presentation.theme.AppThemePreview
-import com.cyrillrx.rpg.core.presentation.theme.borderAlpha
-import com.cyrillrx.rpg.core.presentation.theme.borderWidth
 import com.cyrillrx.rpg.core.presentation.theme.spacingCommon
 import com.cyrillrx.rpg.core.presentation.theme.spacingMedium
 import com.cyrillrx.rpg.core.presentation.theme.spacingSmall
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
 import org.jetbrains.compose.resources.getString
-import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import rpg_companion.composeapp.generated.resources.Res
-import rpg_companion.composeapp.generated.resources.btn_new_character
-import rpg_companion.composeapp.generated.resources.btn_new_character_subtitle
-import rpg_companion.composeapp.generated.resources.btn_quick_create
-import rpg_companion.composeapp.generated.resources.btn_quick_create_subtitle
 import rpg_companion.composeapp.generated.resources.snackbar_character_deleted
 import rpg_companion.composeapp.generated.resources.snackbar_error_deleting_character
 import rpg_companion.composeapp.generated.resources.title_character_list
@@ -143,30 +124,11 @@ fun CharacterListScreen(
                     .padding(paddingValues)
                     .fillMaxSize(),
         ) {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(spacingMedium),
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(spacingCommon),
-            ) {
-                CharacterActionCard(
-                    icon = Icons.Filled.Add,
-                    title = stringResource(Res.string.btn_new_character),
-                    subtitle = stringResource(Res.string.btn_new_character_subtitle),
-                    onClick = onNewCharacterClicked,
-                    filled = true,
-                    modifier = Modifier.weight(1f),
-                )
-                CharacterActionCard(
-                    icon = Icons.Filled.Bolt,
-                    title = stringResource(Res.string.btn_quick_create),
-                    subtitle = stringResource(Res.string.btn_quick_create_subtitle),
-                    onClick = onQuickCreateClicked,
-                    filled = false,
-                    modifier = Modifier.weight(1f),
-                )
-            }
+            CharacterCreateActions(
+                onNewCharacterClicked = onNewCharacterClicked,
+                onQuickCreateClicked = onQuickCreateClicked,
+                modifier = Modifier.padding(spacingCommon),
+            )
 
             Column(modifier = Modifier.weight(1f).fillMaxWidth()) {
                 when (val body = state.body) {
@@ -181,56 +143,6 @@ fun CharacterListScreen(
                         )
                 }
             }
-        }
-    }
-}
-
-@Composable
-private fun CharacterActionCard(
-    icon: ImageVector,
-    title: String,
-    subtitle: String,
-    onClick: () -> Unit,
-    filled: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    val containerColor = if (filled) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.surface
-    val contentColor = if (filled) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurface
-
-    Card(
-        onClick = onClick,
-        shape = MaterialTheme.shapes.medium,
-        border = if (filled) {
-            null
-        } else {
-            BorderStroke(
-                width = borderWidth,
-                color = MaterialTheme.colorScheme.outline.copy(alpha = borderAlpha),
-            )
-        },
-        colors = CardDefaults.cardColors(containerColor = containerColor),
-        modifier = modifier,
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(spacingSmall),
-            modifier = Modifier.padding(spacingCommon),
-        ) {
-            Icon(
-                imageVector = icon,
-                contentDescription = null,
-                tint = contentColor,
-            )
-            Text(
-                text = title,
-                style = MaterialTheme.typography.titleSmall,
-                fontWeight = FontWeight.Bold,
-                color = contentColor,
-            )
-            Text(
-                text = subtitle,
-                style = MaterialTheme.typography.bodySmall,
-                color = contentColor,
-            )
         }
     }
 }
